Add files via upload
Here is the dateset, python code and models to insert, and report.  Please note that the python graphs on the bottom of the python code are designed to be used with two independent variables.  To run the code with more variables, do not select beyond the confusion matrix!

Have fun!